Update nomad v0.9.0

+++ build.sh
@@ -1,8 +1,8 @@
 #!/bin/bash -e
-go get github.com/goreleaser/goreleaser
-git clone https://github.com/nicholasjackson/nomad-arm6l.git
-
-cp ./nomad-arm6l/.goreleaser.yml $GOPATH/src/github.com/hashicorp/nomad
+if ! [[ -f "$GOPATH/src/github.com/hashicorp/nomad/.goreleaser.yml" ]]; then
+    git clone https://github.com/nicholasjackson/nomad-arm6l.git
+    cp ./nomad-arm6l/.goreleaser.yml $GOPATH/src/github.com/hashicorp/nomad
+fi
 
 cd $GOPATH/src/github.com/hashicorp/nomad
-goreleaser --skip-validate --rm-dist --skip-publish  
+goreleaser --skip-validate --rm-dist --skip-publish
